api/requirements.txt
pillow==8.40